Applying suggestions
