+++ libs/libcommon/src/libcommon/config.py
@@ -39,7 +39,7 @@ class AssetsConfig:
 COMMON_HF_ENDPOINT = "https://huggingface.co"
 COMMON_HF_TOKEN = None
 COMMON_LOG_LEVEL = logging.INFO
-COMMON_CONTENT_MAX_SIZE = 1_0000_000
+COMMON_CONTENT_MAX_SIZE = 10_0000_000
 
 
 @dataclass
@@ -60,7 +60,7 @@ class CommonConfig:
                 hf_endpoint=env.str(name="HF_ENDPOINT", default=COMMON_HF_ENDPOINT),
                 hf_token=env.str(name="HF_TOKEN", default=COMMON_HF_TOKEN),  # nosec
                 log_level=env.log_level(name="LOG_LEVEL", default=COMMON_LOG_LEVEL),
-                content_max_size=env.int(name="DOCUMENT_MAX_SIZE", default=COMMON_CONTENT_MAX_SIZE),
+                content_max_size=env.int(name="CONTENT_MAX_SIZE", default=COMMON_CONTENT_MAX_SIZE),
             )
